Rename: Folder.js를 컴포넌트 폴더로 이동
