+++ src/index.js
@@ -1,6 +1,6 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
-import Folder from "Folder";
+import Folder from "components/Folder";
 
 const root = ReactDOM.createRoot(document.getElementById("root"));
 root.render(
